refactor(tools): 分离utils包
